requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiohttp
python-dotenv
openai